requirements.txt
# NRAIZES - Required Python packages
# Install with: pip install -r requirements.txt

# Web Framework
Flask>=3.1.0,<4.0.0
Werkzeug>=3.1.0,<4.0.0
Jinja2>=3.1.0,<4.0.0

# HTTP Client
requests>=2.32.0,<3.0.0
urllib3>=2.5.0,<3.0.0

# Google AI & APIs
google-generativeai>=0.8.0,<1.0.0
google-analytics-admin>=0.27.0,<1.0.0
google-analytics-data>=0.20.0,<1.0.0
google-api-python-client>=2.100.0,<3.0.0
google-auth>=2.40.0,<3.0.0

# Data Processing
pandas>=2.3.0,<3.0.0
numpy>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0

# Environment & Configuration
python-dotenv>=1.2.0,<2.0.0

# CLI
click>=8.3.0,<9.0.0
tabulate>=0.9.0,<1.0.0
tqdm>=4.67.0,<5.0.0

# HTML Parsing
beautifulsoup4>=4.14.0,<5.0.0

# Data Validation
pydantic>=2.12.0,<3.0.0

# Development
colorama>=0.4.6,<1.0.0
